refactor(speakers): update speaker list and detail UI
