+++ src/app/pages/speaker-detail/speaker-detail.scss
@@ -6,3 +6,9 @@
 .speaker-detail p {
   color: #60646b;
 }
+
+.speaker-detail .speaker-icons ion-button {
+  --border-radius: 50%;
+
+  height: 48px;
+}
