+++ src/app/pages/speaker-detail/speaker-detail.ts
@@ -1,6 +1,10 @@
-import { Component, ViewEncapsulation } from '@angular/core';
-import { ActivatedRoute, Router } from '@angular/router';
+import { Component } from '@angular/core';
+import { ActivatedRoute } from '@angular/router';
 import { ConferenceData } from '../../providers/conference-data';
+
+import { InAppBrowser } from '@ionic-native/in-app-browser/ngx';
+
+import { ActionSheetController, AlertController } from '@ionic/angular';
 
 @Component({
   selector: 'page-speaker-detail',
@@ -11,8 +15,10 @@ export class SpeakerDetailPage {
   speaker: any;
 
   constructor(
+    public actionSheetCtrl: ActionSheetController,
+    public alertCtrl: AlertController,
     private dataProvider: ConferenceData,
-    private router: Router,
+    public inAppBrowser: InAppBrowser,
     private route: ActivatedRoute
   ) {}
 
@@ -28,5 +34,71 @@ export class SpeakerDetailPage {
         }
       }
     });
+  }
+
+  openSpeakerSocial(domain: string, speaker: any) {
+    this.inAppBrowser.create(
+      `https://${domain}.com/${speaker.twitter}`,
+      '_blank'
+    );
+  }
+
+  async openSpeakerShare(speaker: any) {
+    const actionSheet = await this.actionSheetCtrl.create({
+      header: 'Share ' + speaker.name,
+      buttons: [
+        {
+          text: 'Copy Link',
+          handler: () => {
+            console.log(
+              'Copy link clicked on https://twitter.com/' + speaker.twitter
+            );
+            if (
+              (window as any)['cordova'] &&
+              (window as any)['cordova'].plugins.clipboard
+            ) {
+              (window as any)['cordova'].plugins.clipboard.copy(
+                'https://twitter.com/' + speaker.twitter
+              );
+            }
+          }
+        },
+        {
+          text: 'Share via ...'
+        },
+        {
+          text: 'Cancel',
+          role: 'cancel'
+        }
+      ]
+    });
+
+    await actionSheet.present();
+  }
+
+  async openContact(speaker: any) {
+    const mode = 'ios'; // this.config.get('mode');
+
+    const actionSheet = await this.actionSheetCtrl.create({
+      header: 'Contact ' + speaker.name,
+      buttons: [
+        {
+          text: `Email ( ${speaker.email} )`,
+          icon: mode !== 'ios' ? 'mail' : null,
+          handler: () => {
+            window.open('mailto:' + speaker.email);
+          }
+        },
+        {
+          text: `Call ( ${speaker.phone} )`,
+          icon: mode !== 'ios' ? 'call' : null,
+          handler: () => {
+            window.open('tel:' + speaker.phone);
+          }
+        }
+      ]
+    });
+
+    await actionSheet.present();
   }
 }
