+++ src/app/pages/speaker-list/speaker-list.html
@@ -11,50 +11,15 @@
   <ion-list>
     <ion-grid fixed>
       <ion-row class="ion-align-items-stretch">
-        <ion-col size="12" size-md="6" *ngFor="let speaker of speakers">
-          <ion-card class="speaker-card">
-            <ion-card-header>
-              <ion-item detail="false" lines="none" routerLink="/app/tabs/speakers/speaker-details/{{speaker.id}}">
-                <ion-avatar slot="start">
-                  <img [src]="speaker.profilePic" alt="Speaker profile pic">
-                </ion-avatar>
-                {{speaker.name}}
-              </ion-item>
+        <ion-col size="12" size-md="6" size-lg="4" *ngFor="let speaker of speakers">
+
+          <ion-card routerLink="/app/tabs/speakers/speaker-details/{{speaker.id}}" class="speaker-card">
+            <ion-card-header translucent>
+              <h1>{{speaker.name}}</h1>
             </ion-card-header>
-
-            <ion-card-content>
-              <ion-list>
-                <ion-item *ngFor="let session of speaker.sessions" routerLink="/app/tabs/speakers/session/{{session.id}}">
-                  <h3>{{session.name}}</h3>
-                </ion-item>
-
-                <ion-item routerLink="/app/tabs/speakers/speaker-details/{{speaker.id}}">
-                  <h3>About {{speaker.name}}</h3>
-                </ion-item>
-              </ion-list>
-            </ion-card-content>
-
-            <ion-row class="ion-no-padding ion-justify-content-center">
-              <ion-col size="4" class="ion-text-left">
-                <ion-button fill="clear" size="small" color="primary" (click)="goToSpeakerTwitter(speaker)">
-                  <ion-icon name="logo-twitter" slot="start"></ion-icon>
-                  Tweet
-                </ion-button>
-              </ion-col>
-              <ion-col size="4" class="ion-text-center">
-                <ion-button fill="clear" size="small" color="primary" (click)="openSpeakerShare(speaker)">
-                  <ion-icon name="share-alt" slot="start"></ion-icon>
-                  Share
-                </ion-button>
-              </ion-col>
-              <ion-col size="4" lass="ion-text-right">
-                <ion-button fill="clear" size="small" color="primary" (click)="openContact(speaker)">
-                  <ion-icon name="chatboxes" slot="start"></ion-icon>
-                  Contact
-                </ion-button>
-              </ion-col>
-            </ion-row>
+            <img [src]="speaker.profilePic" alt="Speaker Profile Picture">
           </ion-card>
+
         </ion-col>
       </ion-row>
     </ion-grid>
