+++ src/app/pages/speaker-list/speaker-list.scss
@@ -2,22 +2,15 @@
   background: #ededed;
 }
 
-.speaker-card {
-  height: 100%;
-  display: flex;
-  flex-direction: column;
-}
-
 .speaker-card ion-card-header {
-  padding: 0;
-}
+  padding-top: 0;
+  padding-bottom: 0;
 
-.speaker-card ion-card-header .item {
-  padding: 4px 16px;
-}
+  color: white;
+  background: rgba(var(--ion-text-color-rgb, 0, 0, 0), 0.4);
 
-.speaker-card ion-card-content {
-  flex: 1 1 auto;
-
-  padding: 0;
+  position: absolute;
+  bottom: 0;
+  left: 0;
+  right: 0;
 }
