+++ src/assets/data/data.json
@@ -248,7 +248,9 @@
     {
       "name": "Burt Bear",
       "profilePic": "/assets/img/speakers/bear.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Burt is a Bear.",
       "location": "Everywhere",
       "email": "burt@example.com",
@@ -258,7 +260,9 @@
     {
       "name": "Charlie Cheetah",
       "profilePic": "/assets/img/speakers/cheetah.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Charlie is a Cheetah.",
       "location": "Everywhere",
       "email": "charlie@example.com",
@@ -268,7 +272,9 @@
     {
       "name": "Donald Duck",
       "profilePic": "/assets/img/speakers/duck.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Donald is a Duck.",
       "location": "Everywhere",
       "email": "donald@example.com",
@@ -278,7 +284,9 @@
     {
       "name": "Eva Eagle",
       "profilePic": "/assets/img/speakers/eagle.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Eva is an Eagle.",
       "location": "Everywhere",
       "email": "eva@example.com",
@@ -288,7 +296,9 @@
     {
       "name": "Ellie Elephant",
       "profilePic": "/assets/img/speakers/elephant.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Ellie is an Elephant.",
       "location": "Everywhere",
       "email": "ellie@example.com",
@@ -298,7 +308,9 @@
     {
       "name": "Gino Giraffe",
       "profilePic": "/assets/img/speakers/giraffe.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Gino is a Giraffe.",
       "location": "Everywhere",
       "email": "gino@example.com",
@@ -308,7 +320,9 @@
     {
       "name": "Isabella Iguana",
       "profilePic": "/assets/img/speakers/iguana.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Isabella is an Iguana.",
       "location": "Everywhere",
       "email": "isabella@example.com",
@@ -318,7 +332,9 @@
     {
       "name": "Karl Kitten",
       "profilePic": "/assets/img/speakers/kitten.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Karl is a Kitten.",
       "location": "Everywhere",
       "email": "karl@example.com",
@@ -328,7 +344,9 @@
     {
       "name": "Lionel Lion",
       "profilePic": "/assets/img/speakers/lion.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Lionel is a Lion.",
       "location": "Everywhere",
       "email": "lionel@example.com",
@@ -338,7 +356,9 @@
     {
       "name": "Molly Mouse",
       "profilePic": "/assets/img/speakers/mouse.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Molly is a Mouse.",
       "location": "Everywhere",
       "email": "molly@example.com",
@@ -348,7 +368,9 @@
     {
       "name": "Paul Puppy",
       "profilePic": "/assets/img/speakers/puppy.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Paul is a Puppy.",
       "location": "Everywhere",
       "email": "paul@example.com",
@@ -358,7 +380,9 @@
     {
       "name": "Rachel Rabbit",
       "profilePic": "/assets/img/speakers/rabbit.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Rachel is a Rabbit.",
       "location": "Everywhere",
       "email": "rachel@example.com",
@@ -368,7 +392,9 @@
     {
       "name": "Ted Turtle",
       "profilePic": "/assets/img/speakers/turtle.jpg",
+      "instagram": "ionicframework",
       "twitter": "ionicframework",
+      "github": "ionic-team/ionic",
       "about": "Ted is a Turtle.",
       "location": "Everywhere",
       "email": "ted@example.com",
